+++ CHANGELOG.md
@@ -6,7 +6,9 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Added
 
-- Nothing.
+- [#24](https://github.com/phly/keep-a-changelog/pull/24) adds a GitLab repository provider.
+  You can specify the new provider via a new `--provider gitlab` option to
+  either the `tag` or `release` commands.
 
 ### Changed